+++ src/main/kotlin/com/github/rafaelldi/tyeplugin/settings/TyeSettingsConfigurable.kt
@@ -1,34 +1,57 @@
 package com.github.rafaelldi.tyeplugin.settings
 
-import com.github.rafaelldi.tyeplugin.listeners.TyeGlobalToolListener
+import com.github.rafaelldi.tyeplugin.services.TyeCliService
 import com.github.rafaelldi.tyeplugin.services.TyePathProvider.Companion.TYE_TOOL
+import com.intellij.openapi.application.invokeLater
+import com.intellij.openapi.components.service
 import com.intellij.openapi.fileChooser.FileChooserDescriptor
+import com.intellij.openapi.fileChooser.FileChooserDescriptorFactory
 import com.intellij.openapi.options.BoundConfigurable
+import com.intellij.openapi.progress.ProgressIndicator
+import com.intellij.openapi.progress.ProgressManager
+import com.intellij.openapi.progress.Task
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.DialogPanel
-import com.intellij.ui.layout.panel
-import com.intellij.util.messages.MessageBus
+import com.intellij.openapi.ui.TextFieldWithBrowseButton
+import com.intellij.openapi.util.NlsContexts.DialogTitle
+import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.ui.components.textFieldWithBrowseButton
+import com.intellij.ui.layout.*
+import java.io.File
+import javax.swing.JTextField
 
 class TyeSettingsConfigurable(private val project: Project) : BoundConfigurable("Tye") {
-    private val settings: TyeSettingsState = TyeSettingsState.getInstance(project)
-    private val messageBus: MessageBus = project.messageBus
+    private val settings: TyeSettings get() = TyeSettings.getInstance(project)
+    private val tyeCliService: TyeCliService = project.service()
 
     override fun createPanel(): DialogPanel = panel {
         row("Tye tool path") {
-            textFieldWithBrowseButton(
-                settings::tyeToolPath,
+            val pathTextField = JBTextFieldWithSecondaryValue()
+            textFieldWithBrowseButton(settings::tyeToolPath.toNullableBinding(""),
                 "Select Path",
+                pathTextField as JTextField,
                 project,
-                FileChooserDescriptor(true, false, false, false, false, false)
-                    .withFileFilter { vf -> vf.nameWithoutExtension.lowercase() == TYE_TOOL }
-                    .withTitle("Select Path")
-            ) { chosenFile ->
-                messageBus.syncPublisher(TyeGlobalToolListener.TOPIC).tyeToolPathChanged(chosenFile.path)
+                FileChooserDescriptorFactory.createSingleFileNoJarsDescriptor().withFileFilter {
+                    it.nameWithoutExtension.lowercase() == TYE_TOOL
+                })
+            { chosenFile ->
+                ProgressManager.getInstance().run(object : Task.Backgroundable(project, "Get tye version") {
+                    override fun run(indicator: ProgressIndicator) {
+                        val version = tyeCliService.getVersion(chosenFile.path)
+                        if (version != null) {
+                            invokeLater {
+                                pathTextField.setSecondaryValue(version.toString())
+                            }
+                        }
+                    }
+                })
                 chosenFile.path
             }
-        }
-        row("Tye version") {
-            textField(settings::tyeToolVersion).enabled(false)
+                .withValidationOnInput {
+                    val file = File(it.text)
+                    if (file.exists() && file.isDirectory.not() && file.canExecute()) return@withValidationOnInput null
+                    return@withValidationOnInput error("Invalid path for tye executable")
+                }
         }
         titledRow("Options") {
             row {
@@ -39,4 +62,20 @@ class TyeSettingsConfigurable(private val project: Project) : BoundConfigurable(
             }
         }
     }
+}
+
+
+fun Cell.textFieldWithBrowseButton(
+    modelBinding: PropertyBinding<String>,
+    @DialogTitle browseDialogTitle: String,
+    textField: JTextField,
+    project: Project? = null,
+    fileChooserDescriptor: FileChooserDescriptor = FileChooserDescriptorFactory.createSingleFileNoJarsDescriptor(),
+    fileChosen: ((chosenFile: VirtualFile) -> String)? = null
+): CellBuilder<TextFieldWithBrowseButton> {
+    val textFieldWithBrowseButton =
+        textFieldWithBrowseButton(project, browseDialogTitle, textField, fileChooserDescriptor, fileChosen)
+    textFieldWithBrowseButton.text = modelBinding.get()
+    return component(textFieldWithBrowseButton).constraints(growX)
+        .withBinding(TextFieldWithBrowseButton::getText, TextFieldWithBrowseButton::setText, modelBinding)
 }