+++ src/main/kotlin/com/github/rafaelldi/tyeplugin/settings/TyeSettingsState.kt
@@ -1,29 +1,10 @@
 package com.github.rafaelldi.tyeplugin.settings
 
 import com.github.rafaelldi.tyeplugin.services.TyePathProvider
-import com.intellij.openapi.components.PersistentStateComponent
-import com.intellij.openapi.components.State
-import com.intellij.openapi.components.Storage
-import com.intellij.openapi.components.service
-import com.intellij.openapi.project.Project
-import com.intellij.util.xmlb.XmlSerializerUtil
+import com.intellij.openapi.components.BaseState
 
-@State(
-    name = "com.github.rafaelldi.tyeplugin.settings.TyeSettingsState",
-    storages = [(Storage("TyePluginSettings.xml"))]
-)
-class TyeSettingsState : PersistentStateComponent<TyeSettingsState> {
-    companion object {
-        @JvmStatic
-        fun getInstance(project: Project): TyeSettingsState = project.service()
-    }
-
-    var tyeToolPath: String = TyePathProvider.getDefaultGlobalPath()
-    var tyeToolVersion: String = ""
-    var overwriteTyeFile = false
-    var checkTyeNewVersions = false
-
-    override fun getState(): TyeSettingsState = this
-
-    override fun loadState(state: TyeSettingsState) = XmlSerializerUtil.copyBean(state, this)
+class TyeSettingsState : BaseState() {
+    var tyeToolPath by string(TyePathProvider.getDefaultGlobalPath())
+    var overwriteTyeFile by property(false)
+    var checkTyeNewVersions by property(false)
 }